requirements.txt
Django==2.2.2
Pillow==6.0.0
psycopg2-binary==2.8.2
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
